// src/CommandBlock.Frontend/src/app/api/model/modpackSearchResult.ts
/**
 * CommandBlock.API | v1
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { BackupEntryDtoSizeBytes } from './backupEntryDtoSizeBytes';


export interface ModpackSearchResult { 
    slug: string;
    title: string;
    description: string;
    iconUrl: string | null;
    downloads: BackupEntryDtoSizeBytes;
}